Use the SQL import handler to test the user account export
OX-12522

This change was necessary since the tests do not use the Mongo Tenants
framework and disabling it would break other tests.
