+++ src/test/resources/test.conf
@@ -3,7 +3,7 @@ docker.file = ["src/test/resources/docker-compose.yml"]
 sirius.frameworks {
     biz.tenants = true
     biz.tenants-jdbc = true
-    biz.tenants-mongo = true
+    biz.tenants-mongo = false
     biz.code-lists-jdbc = true
     biz.code-lists-mongo = false
     biz.protocols = true
